Add :change-idea transform (incomplete).
Now I understand why you'd add the basic (empty) HTML on `:node-create`:
Because then you can use one set of functions to change and create the
data. I'm still a bit sceptic because you'd create the template in the
data first (with bogus data) and only then fill it in. That might be not
so good for performance, but I'm not sure and for now I don't bother.